requirements.txt
# Brain Tumor 3-Class Discrimination System
# Python 3.10+

# Core ML
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Medical Imaging
nibabel>=5.1.0
SimpleITK>=2.3.0
monai>=1.3.0

# BraTS Pre-trained Models (Docker 필요)
brats>=0.0.14

# Data Processing
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
pathlib2>=2.3.7

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0
